requirements.txt
alembic==1.7.5
aniso8601==9.0.1
antiorm==1.2.1
black==21.12b0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cryptography==36.0.0
db==0.1.1
Flask==2.0.2
Flask-HTTPAuth==4.5.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
generate-password==1.0.0
greenlet==1.1.2
importlib-metadata==4.8.2
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
jwt==1.3.1
lark-parser==0.12.0
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
pathspec==0.9.0
permission==0.4.1
platformdirs==2.4.0
psycopg2==2.9.2
psycopg2-binary==2.9.2
pycparser==2.21
PyJWT==2.3.0
python-decouple==3.5
python-dotenv==0.19.2
pytz==2021.3
required==0.4.0
six==1.16.0
SQLAlchemy==1.4.27
tomli==1.2.2
typed-ast==1.5.1
typing-extensions==4.0.0
Werkzeug==2.0.2
zipp==3.6.0
